Gather registration in more or less one file.
