+++ src/coreComponents/python/modules/geosx_mesh_doctor/checks/__init__.py
@@ -1,6 +1,4 @@
-COLLOCATES_NODES = "collocated_nodes"
-ELEMENT_VOLUMES = "element_volumes"
-GENERATE_FRACTURES = "generate_fractures"
-GENERATE_GLOBAL_IDS = "generate_global_ids"
+from typing import Dict, Any
 
-all_checks = dict()
+# Singleton-like pattern for checks registration
+all_checks: Dict[str, Any] = dict()
