+++ src/coreComponents/python/modules/geosx_mesh_doctor/mesh_doctor.py
@@ -2,9 +2,9 @@ import logging
 import sys
 
 from checks import all_checks
-from parsing import register
-from parsing import all_checks_helpers
+from parsing import all_checks_helpers, CheckHelper
 from parsing.cli_parsing import parse, parse_and_set_verbosity
+import register
 
 
 def main():
@@ -22,7 +22,7 @@ def main():
         except KeyError as e:
             logging.critical(f"Check {check_name} is not a valid check.")
             sys.exit(1)
-        helper = all_checks_helpers[check_name]
+        helper: CheckHelper = all_checks_helpers[check_name]
         result = check(args.vtk_input_file, check_options)
         helper.display_results(check_options, result)
