+++ src/coreComponents/python/modules/geosx_mesh_doctor/parsing/__init__.py
@@ -1,5 +1,11 @@
 from dataclasses import dataclass
-from typing import Callable, Any
+from typing import Dict, Callable, Any
+
+
+COLLOCATES_NODES = "collocated_nodes"
+ELEMENT_VOLUMES = "element_volumes"
+GENERATE_FRACTURES = "generate_fractures"
+GENERATE_GLOBAL_IDS = "generate_global_ids"
 
 
 @dataclass(frozen=True)
@@ -9,16 +15,5 @@ class CheckHelper:
     get_help: Callable[[None], str]
 
 
-def build_check_helper(module) -> CheckHelper:
-    """
-    If a module has functions `parse_cli_options`, `display_results` and `get_help`,
-    the `CheckHelper` built from those functions.
-    :param module: Any module
-    :return: The CheckHelper instance.
-    """
-    return CheckHelper(parse_cli_options=module.parse_cli_options,
-                       display_results=module.display_results,
-                       get_help=module.get_help)
-
-
-all_checks_helpers = dict()
+# Singleton-like pattern for check helpers registration
+all_checks_helpers: Dict[str, CheckHelper] = dict()
